majit-translate: delete dead syn_metadata helpers and relocate syn-free utils
Delete six syn_metadata helpers that have no remaining callers (the
import-aware type-string machinery that served the removed AST
`Expr::Path` resolver): full_type_string,
qualified_full_type_string_with_imports, collect_trait_names,
extract_dyn_trait_root (+ _with_context), qualify_known_trait_name,
trait_object_root_name_qualified.

Move the syn-free type-id string classifiers (transparent_result_ok_type,
first_top_level_generic_arg, nolength_from_array_type_id) into a new
front::typestr module, and move the unit-variant ctor allowlist
(is_synthetic_unit_variant_path) into translator::rtyper::unit_variant_fold
next to its fold. Narrow trait_object_root_name to pub(crate); its only
caller is type_root_ident.

front::syn_metadata now holds only the four syn-tree harvesters the MIR
path still sources from interpreter source: collect_struct_origins,
classify_fn_arg_ty, type_root_ident, trait_object_root_name.
