+++ majit/majit-translate/src/front/mod.rs
@@ -74,5 +74,6 @@ pub mod opcode_wrapper;
 pub mod raise;
 pub mod semantic;
 pub mod syn_metadata;
+pub mod typestr;
 
 pub use semantic::{AstGraphOptions, SemanticFunction, SemanticProgram, StructFieldRegistry};
