+++ majit/majit-translate/src/front/typestr.rs
@@ -1,0 +1,89 @@
+//! Syn-free type-id string classifiers used on the MIR path.
+//!
+//! These helpers inspect type identities as plain strings (the
+//! `array_type_id` / declared-signature strings the codewriter already
+//! carries); they parse no `syn` tree and depend on nothing in
+//! [`crate::front::syn_metadata`].  Their only consumers live in
+//! `jit_codewriter::call`.
+
+/// Detect the canonical `Result<T, …>` wrapper and project the inner
+/// `T`.  Returns `None` for non-`Result` shapes, for `Result<(), …>`
+/// (no transparent type to project), and for malformed inputs.
+///
+/// The only consumers live in `jit_codewriter::call`.
+pub fn transparent_result_ok_type(type_str: &str) -> Option<&str> {
+    let trimmed = type_str.trim();
+    for prefix in ["Result<", "std::result::Result<", "core::result::Result<"] {
+        let Some(inner) = trimmed
+            .strip_prefix(prefix)
+            .and_then(|rest| rest.strip_suffix('>'))
+        else {
+            continue;
+        };
+        let ok_type = first_top_level_generic_arg(inner).map(str::trim)?;
+        if ok_type == "()" {
+            return None;
+        }
+        return Some(ok_type);
+    }
+    None
+}
+
+/// Return the first comma-delimited top-level generic argument in
+/// `args` (`"A, B<C, D>, E"` → `"A"`).  Tracks bracket depth so a
+/// nested generic boundary does not confuse the split.
+///
+/// Used by [`transparent_result_ok_type`].
+pub fn first_top_level_generic_arg(args: &str) -> Option<&str> {
+    let mut depth = 0usize;
+    for (idx, ch) in args.char_indices() {
+        match ch {
+            '<' | '(' | '[' => depth += 1,
+            '>' | ')' | ']' => depth = depth.saturating_sub(1),
+            ',' if depth == 0 => return Some(&args[..idx]),
+            _ => {}
+        }
+    }
+    if args.is_empty() { None } else { Some(args) }
+}
+
+/// Decide whether a registered `array_type_id` describes a
+/// headerless item-run pointee or a length-prefixed wrapper.  Bare
+/// pointers to identifier types address `items[0]` (no length word);
+/// `Vec<T>` / `GcArray<T>` / `Ptr(GcArray(T))` shapes carry a length
+/// header at offset 0 and therefore keep the PyPy default `False`.
+///
+/// Only `jit_codewriter::call` consumes it.
+pub fn nolength_from_array_type_id(array_type_id: Option<&str>) -> bool {
+    let Some(s) = array_type_id else {
+        return false;
+    };
+    let mut inner = s.trim();
+    loop {
+        let stripped = inner
+            .strip_prefix("*const ")
+            .or_else(|| inner.strip_prefix("*mut "))
+            .or_else(|| inner.strip_prefix("&mut "))
+            .or_else(|| inner.strip_prefix('&'));
+        match stripped {
+            Some(rest) => inner = rest.trim_start(),
+            None => break,
+        }
+    }
+    if inner.starts_with('[') && inner.ends_with(']') {
+        return true;
+    }
+    // Length-prefixed wrappers carry `<` (generic) or `(` (paren-style
+    // lltype spelling such as `Ptr(GcArray(...))`).  Keep the PyPy
+    // default `False` for those — a pointer to a wrapper still
+    // dereferences a length header.
+    if inner.contains('<') || inner.contains('(') {
+        return false;
+    }
+    // Bare identifier pointee (`*const i64`, `*const Point`) means the
+    // pointer addresses items[0] of a primitive / struct item type.
+    // A bare identifier with NO pointer prefix is a value-type binding
+    // (e.g. an `array_type_id` directly naming a struct that contains
+    // an embedded array); preserve the PyPy default `False` for that.
+    s.trim() != inner
+}
